requirements.txt
pandas
openpyxl
fpdf
yagmail
python-dotenv